Move `generator` tests into `core`
