+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/testutil/TestHelpers.kt
@@ -7,20 +7,25 @@ package software.amazon.smithy.rust.codegen.core.testutil
 
 import software.amazon.smithy.model.Model
 import software.amazon.smithy.model.knowledge.NullableIndex
+import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.shapes.StructureShape
 import software.amazon.smithy.rust.codegen.core.rustlang.Attribute
 import software.amazon.smithy.rust.codegen.core.rustlang.CargoDependency
 import software.amazon.smithy.rust.codegen.core.rustlang.CratesIo
 import software.amazon.smithy.rust.codegen.core.rustlang.DependencyScope
+import software.amazon.smithy.rust.codegen.core.rustlang.RustReservedWordSymbolProvider
 import software.amazon.smithy.rust.codegen.core.rustlang.RustWriter
 import software.amazon.smithy.rust.codegen.core.rustlang.asType
+import software.amazon.smithy.rust.codegen.core.smithy.BaseSymbolMetadataProvider
+import software.amazon.smithy.rust.codegen.core.smithy.CodegenContext
 import software.amazon.smithy.rust.codegen.core.smithy.CodegenTarget
 import software.amazon.smithy.rust.codegen.core.smithy.CoreCodegenConfig
 import software.amazon.smithy.rust.codegen.core.smithy.CoreRustSettings
 import software.amazon.smithy.rust.codegen.core.smithy.RuntimeConfig
 import software.amazon.smithy.rust.codegen.core.smithy.RuntimeCrateLocation
 import software.amazon.smithy.rust.codegen.core.smithy.RustSymbolProvider
+import software.amazon.smithy.rust.codegen.core.smithy.SymbolVisitor
 import software.amazon.smithy.rust.codegen.core.smithy.SymbolVisitorConfig
 import software.amazon.smithy.rust.codegen.core.smithy.generators.BuilderGenerator
 import software.amazon.smithy.rust.codegen.core.smithy.generators.StructureGenerator
@@ -68,6 +73,31 @@ fun String.asSmithyModel(sourceLocation: String? = null, smithyVersion: String =
     return Model.assembler().discoverModels().addUnparsedModel(sourceLocation ?: "test.smithy", processed).assemble()
         .unwrap()
 }
+
+// Intentionally only visible to codegen-core since the other modules have their own symbol providers
+internal fun testSymbolProvider(model: Model): RustSymbolProvider = SymbolVisitor(
+    model,
+    ServiceShape.builder().version("test").id("test#Service").build(),
+    TestSymbolVisitorConfig,
+).let { BaseSymbolMetadataProvider(it, model, additionalAttributes = listOf(Attribute.NonExhaustive)) }
+    .let { RustReservedWordSymbolProvider(it, model) }
+
+// Intentionally only visible to codegen-core since the other modules have their own contexts
+internal fun testCodegenContext(
+    model: Model,
+    serviceShape: ServiceShape? = null,
+    settings: CoreRustSettings = testRustSettings(),
+    codegenTarget: CodegenTarget = CodegenTarget.CLIENT,
+): CodegenContext = CodegenContext(
+    model,
+    testSymbolProvider(model),
+    serviceShape
+        ?: model.serviceShapes.firstOrNull()
+        ?: ServiceShape.builder().version("test").id("test#Service").build(),
+    ShapeId.from("test#Protocol"),
+    settings,
+    codegenTarget,
+)
 
 /**
  * In tests, we frequently need to generate a struct, a builder, and an impl block to access said builder.
